+++ CHANGELOG.md
@@ -7,6 +7,16 @@ The major version bump is due to dropped support for PHP `8.0` and a several bre
 * Added `Session::disableAutoCommit()` method.
 * Added `Session::enableAutoCommit()` method.
 * Made the `Session::gc()` method public and added a `$force` parameter.
+* The following `Request` properties are now public readonly:
+	- `Request::$query`
+	- `Request::$post`
+	- `Request::$cookies`
+	- `Request::$files`
+	- `Request::$server`
+	- `Request::$headers`
+* The following `Response` properties are now public readonly:
+	- `Response::$cookies`
+	- `Response::$headers`
 
 #### Changes